bug fix for deleting authors

+++ code/socialDistribution/fetchers.py
@@ -55,10 +55,11 @@ def update_authors_for_server(server_url):
 
         # check for deleted authors 
         for author in Author.objects.all():
-            author_url = author.url.strip("/")
-            res_code, res_data = api_requests.get(author_url)
-            if res_code == 404 or res_code == 410:
-                author.delete()
+            if not author.up_to_date():
+                author_url = author.url.strip("/")
+                res_code, res_data = api_requests.get(author_url)
+                if res_code == 404 or res_code == 410:
+                    author.delete()
 
     except Exception as e:
         logger.error(e, exc_info=True)
